Rahul's changes and addition of favicon to all the pages

+++ public/js/dashboard.js
@@ -13,6 +13,7 @@ let scanButtonClicked = false;
 let deviceListArray = [];
 let tabStatus = 'current';
 let webSocket;
+let simulationStatus;
 
 
 document.getElementById('defaultOpen').addEventListener('click', () => {
@@ -27,6 +28,7 @@ document.getElementById('HistoricTab').addEventListener('click', () => {
 
 document.addEventListener("DOMContentLoaded", async (event) => {
     loading();
+
     //////////////////////////////////CURRENT TAB//////////////////////////
 
     // Get the list of devices
@@ -67,8 +69,9 @@ document.addEventListener("DOMContentLoaded", async (event) => {
     // console.log(responseDeviceID);
     if (responseDeviceID) {
         sessionStorage.setItem('deviceID', responseDeviceID);
-        // getsparklineCurve();
+        getsparklineCurve();
     }
+    getsimulationstatus();
 
     // Set the labelSelector to the label obtained from deviceData.
     const deviceDataSelector = deviceListArray.find(entry => entry.deviceId === sessionStorage.getItem('deviceID'));
@@ -100,9 +103,11 @@ document.addEventListener("DOMContentLoaded", async (event) => {
             if (tabStatus === 'historic') {
                 performHistoricActionsWithDeviceID();
             } else {
+                loading();
+                getsimulationstatus();
                 performActionsWithDeviceID();
             };
-            // getsparklineCurve();
+            getsparklineCurve();
         };
     });
     // Get the checkbox element and if it is already checked, pass a status code 1 to plot both IV and PV curve. If not checked pass status code 0..
@@ -415,14 +420,25 @@ function updateUI(indexValue) {
     createHistoricSparklineImage(jsonData['ff'], 'sparkline-image-historical-ff', `${jsonData['ff'][indexValue].value}`, jsonData['ff'][indexValue].value, historicTs);
 
     // Tcell DATA
-    createHistoricSparklineImage(jsonData['simTemperature'], 'sparkline-image-historical-tcell', `${jsonData['simTemperature'][indexValue].value}°C`, jsonData['simTemperature'][indexValue].value, historicTs);
+    if (simulationStatus === true) {
+        createHistoricSparklineImage(jsonData['simTemperature'], 'sparkline-image-historical-tcell', `${jsonData['simTemperature'][indexValue].value}°C`, jsonData['simTemperature'][indexValue].value, historicTs);
+    } else {
+        createzerosparkline('sparkline-image-historical-tcell')
+    }
 
     // GEFF DATA
-    createHistoricSparklineImage(jsonData['simIrradiance'], 'sparkline-image-historical-Geff', `${jsonData['simIrradiance'][indexValue].value}W/m²`, jsonData['simIrradiance'][indexValue].value, historicTs);
+    if (simulationStatus === true) {
+        createHistoricSparklineImage(jsonData['simIrradiance'], 'sparkline-image-historical-Geff', `${jsonData['simIrradiance'][indexValue].value}W/m²`, jsonData['simIrradiance'][indexValue].value, historicTs);
+    } else {
+        createzerosparkline('sparkline-image-historical-Geff')
+    }
 
     // PF DATA
-    createHistoricSparklineImage(jsonData['performanceFactor'], 'sparkline-image-historical-pf', `${jsonData['performanceFactor'][indexValue].value}%`, jsonData['performanceFactor'][indexValue].value, historicTs);
-
+    if (simulationStatus === true) {
+        createHistoricSparklineImage(jsonData['performanceFactor'], 'sparkline-image-historical-pf', `${jsonData['performanceFactor'][indexValue].value}%`, jsonData['performanceFactor'][indexValue].value, historicTs);
+    } else {
+        createzerosparkline('sparkline-image-historical-pf')
+    }
     // CREATE THE HISTORIC CURVY CHART
     timestampSlider.value = indexValue;
     historiccurrent = jsonData['currents'][indexValue].value;
@@ -620,18 +636,12 @@ async function performActionsWithDeviceID() {
         } else {
             if (parsed_data.data && parsed_data.data.appStatus && parsed_data.data.appStatus[0] && parsed_data.data.appStatus[0][0]) {
                 lasttime = parsed_data.data.appStatus[0][0];
-            } else if (parsed_data.data && parsed_data.data.voc && parsed_data.data.voc[0] && parsed_data.data.voc[0][0]) {
-                lasttime = parsed_data.data.voc[0][0];
-            } else {
-                lasttime = parsed_data.data.performanceFactor[0][0];
+                sessionStorage.setItem('lastscan', lasttime)
             };
-            difflast = ((new Date().getTime()) - lasttime) / 60000;
-            if (difflast < 30) {
-                getsparklineCurve();
-            } else {
+            difflast = ((new Date().getTime()) - (sessionStorage.getItem('lastscan'))) / 60000;
+            if (difflast > 30) {
                 sessionStorage.setItem('trendLineData', JSON.stringify({}));
-            }
-
+            };
             if (parsed_data.data && parsed_data.data.currents && parsed_data.data.currents.length > 0) {
                 sessionStorage.setItem('currentPoints', parsed_data.data.currents[0][1]);
                 sessionStorage.setItem('voc', parsed_data.data.voc[0][1]);
@@ -705,15 +715,19 @@ async function performActionsWithDeviceID() {
                 rssi_lastupdate = parsed_data.data.rssi[0][0];
                 rssi = parsed_data.data.rssi[0][1];
                 updateWifiSignal(rssi, rssi_lastupdate);
-
-
+            }
+            if (simulationStatus === true) {
+                if (parsed_data.data && parsed_data.data.simTemperature && parsed_data.data.simTemperature.length > 0) {
+                    sessionStorage.setItem('simT', parsed_data.data.simTemperature[0][1]);
+                    sessionStorage.setItem('simG', parsed_data.data.simIrradiance[0][1]);
+                    sessionStorage.setItem('pf', parsed_data.data.performanceFactor[0][1]);
+                }
+            } else {
+                sessionStorage.setItem('simT', 0);
+                sessionStorage.setItem('simG', 0);
+                sessionStorage.setItem('pf', 0);
             }
 
-            if (parsed_data.data && parsed_data.data.simTemperature && parsed_data.data.simTemperature.length > 0) {
-                sessionStorage.setItem('simT', parsed_data.data.simTemperature[0][1]);
-                sessionStorage.setItem('simG', parsed_data.data.simIrradiance[0][1]);
-                sessionStorage.setItem('pf', parsed_data.data.performanceFactor[0][1]);
-            };
 
             if (parsed_data.data && parsed_data.data.iSlow && parsed_data.data.iSlow.length > 0) {
                 sessionStorage.setItem('iop', parsed_data.data.iSlow[0][1]);
@@ -775,27 +789,32 @@ async function performActionsWithDeviceID() {
             // Calculate fill factor
             fillfactor = Math.round((voltageAtMaxPoint * currentAtMaxPoint) / (formattedVoc * formattedIsc) * 100) / 100;
 
-            if (parsed_data.data && parsed_data.data.currents && parsed_data.data.currents.length > 0) {
-                if (formattedVoc != 0 || formattedIsc != 0) {
-                    const simulatedCurvy = { pageIdentifier: pageIdentifier, voc: formattedVoc, isc: formattedIsc, deviceID: sessionStorage.getItem('deviceID') };
-                    const simCurvyOptions = {
-                        method: "POST",
-                        headers: {
-                            "Content-Type": "application/json",
-                            'Authorization': `Bearer ${token}`,
-                        },
-                        body: JSON.stringify(simulatedCurvy),
+            if (simulationStatus === true) {
+                if (parsed_data.data && parsed_data.data.currents && parsed_data.data.currents.length > 0) {
+                    if (formattedVoc != 0 || formattedIsc != 0) {
+                        const simulatedCurvy = { pageIdentifier: pageIdentifier, voc: formattedVoc, isc: formattedIsc, deviceID: sessionStorage.getItem('deviceID') };
+                        const simCurvyOptions = {
+                            method: "POST",
+                            headers: {
+                                "Content-Type": "application/json",
+                                'Authorization': `Bearer ${token}`,
+                            },
+                            body: JSON.stringify(simulatedCurvy),
+                        };
+
+                        const responseSimCurvy = await fetch("/deviceAttribute", simCurvyOptions);
+                        const simulatedCurvyData = await responseSimCurvy.json();
+                        sessionStorage.setItem('simCurrent', handleNull(simulatedCurvyData["Current"]));
+                        sessionStorage.setItem('simVoltage', handleNull(simulatedCurvyData["Voltage"]));
+                    } else {
+                        sessionStorage.setItem('simCurrent', []);
+                        sessionStorage.setItem('simVoltage', []);
                     };
 
-                    const responseSimCurvy = await fetch("/deviceAttribute", simCurvyOptions);
-                    const simulatedCurvyData = await responseSimCurvy.json();
-                    sessionStorage.setItem('simCurrent', handleNull(simulatedCurvyData["Current"]));
-                    sessionStorage.setItem('simVoltage', handleNull(simulatedCurvyData["Voltage"]));
-                } else {
-                    sessionStorage.setItem('simCurrent', []);
-                    sessionStorage.setItem('simVoltage', []);
                 };
-
+            } else {
+                sessionStorage.setItem('simCurrent', []);
+                sessionStorage.setItem('simVoltage', []);
             };
 
             simCurrent = sessionStorage.getItem('simCurrent').split(",").map(parseFloat);
@@ -807,26 +826,46 @@ async function performActionsWithDeviceID() {
             checkboxstatuscode = sessionStorage.getItem('checkboxstatus');
             createIVchart(formattedCurrent, voltagepoints, voltageAtMaxPoint, currentAtMaxPoint, powerAtMaxPoint, formattedVoperating, formattedIoperating, formattedOperatingPower, formattedPower, 'chart-curvy', simCurrent, simVoltage, checkboxstatuscode);
 
-            cellTemperature = sessionStorage.getItem('simT');
-            simIrradiance = sessionStorage.getItem('simG');
-            performanceFactor = sessionStorage.getItem('pf');
+            if (simulationStatus === true) {
+                cellTemperature = sessionStorage.getItem('simT');
+                simIrradiance = sessionStorage.getItem('simG');
+                performanceFactor = sessionStorage.getItem('pf');
 
-            // create sparkline chart
-            createSparklineImage('sparkline-image-voc', `${formattedVoc} V`);
-            createSparklineImage('sparkline-image-isc', `${formattedIsc} A`);
-            createSparklineImage('sparkline-image-ff', `${fillfactor}`);
+                // create sparkline chart
+                createSparklineImage('sparkline-image-voc', `${formattedVoc} V`);
+                createSparklineImage('sparkline-image-isc', `${formattedIsc} A`);
+                createSparklineImage('sparkline-image-ff', `${fillfactor}`);
 
-            createSparklineImage('sparkline-image-vmp', `${voltageAtMaxPoint} V`);
-            createSparklineImage('sparkline-image-imp', `${currentAtMaxPoint} A`);
-            createSparklineImage('sparkline-image-pmp', `${powerAtMaxPoint}W`);
+                createSparklineImage('sparkline-image-vmp', `${voltageAtMaxPoint} V`);
+                createSparklineImage('sparkline-image-imp', `${currentAtMaxPoint} A`);
+                createSparklineImage('sparkline-image-pmp', `${powerAtMaxPoint}W`);
 
-            createSparklineImage('sparkline-image-vop', `${formattedVoperating} V`);
-            createSparklineImage('sparkline-image-iop', `${formattedIoperating} A`);
-            createSparklineImage('sparkline-image-pop', `${formattedOperatingPower}W`);
+                createSparklineImage('sparkline-image-vop', `${formattedVoperating} V`);
+                createSparklineImage('sparkline-image-iop', `${formattedIoperating} A`);
+                createSparklineImage('sparkline-image-pop', `${formattedOperatingPower}W`);
 
-            createSparklineImage('sparkline-image-temperature', `${cellTemperature}°C`);
-            createSparklineImage('sparkline-image-irradiance', `${simIrradiance}W/m²`);
-            createSparklineImage('sparkline-image-perfromance-factor', `${performanceFactor}%`);
+                createSparklineImage('sparkline-image-temperature', `${cellTemperature}°C`);
+                createSparklineImage('sparkline-image-irradiance', `${simIrradiance}W/m²`);
+                createSparklineImage('sparkline-image-perfromance-factor', `${performanceFactor}%`);
+            } else {
+                // create sparkline chart
+                createSparklineImage('sparkline-image-voc', `${formattedVoc} V`);
+                createSparklineImage('sparkline-image-isc', `${formattedIsc} A`);
+                createSparklineImage('sparkline-image-ff', `${fillfactor}`);
+
+                createSparklineImage('sparkline-image-vmp', `${voltageAtMaxPoint} V`);
+                createSparklineImage('sparkline-image-imp', `${currentAtMaxPoint} A`);
+                createSparklineImage('sparkline-image-pmp', `${powerAtMaxPoint}W`);
+
+                createSparklineImage('sparkline-image-vop', `${formattedVoperating} V`);
+                createSparklineImage('sparkline-image-iop', `${formattedIoperating} A`);
+                createSparklineImage('sparkline-image-pop', `${formattedOperatingPower}W`);
+
+                createzerosparkline('sparkline-image-temperature')
+                createzerosparkline('sparkline-image-irradiance')
+                createzerosparkline('sparkline-image-perfromance-factor')
+            }
+
         };
     };
 
@@ -1273,7 +1312,7 @@ async function createSparklineImage(elementID, value_sparline) {
                 text: `${value_sparline}`,
                 showarrow: false,
                 font: {
-                    size: 25,
+                    size: 20,
                     color: 'brown',
                 },
             }
@@ -1286,8 +1325,10 @@ async function createSparklineImage(elementID, value_sparline) {
     };
     Plotly.newPlot(elementID, [trace], layout, config);
 
-    Plotly.toImage(elementID, { format: 'png', width: 400, height: 80 }).then(function (url) {
+    Plotly.toImage(elementID, { format: 'svg', width: 250, height: 50 }).then(function (url) {
         document.getElementById(elementID).src = url;
+        document.getElementById(elementID).classList.remove('image-modify2');
+        document.getElementById(elementID).classList.add('sparklineimages');
     });
 };
 
@@ -2196,7 +2237,7 @@ function createzerosparkline(elementID) {
                 text: textval,
                 showarrow: false,
                 font: {
-                    size: 25,
+                    size: 20,
                     color: 'grey',
                 },
             }
@@ -2209,8 +2250,10 @@ function createzerosparkline(elementID) {
     };
     // Plot the chart
     Plotly.newPlot(elementID, data, layout, config);
-    Plotly.toImage(elementID, { format: 'png', width: 250, height: 80 }).then(function (url) {
+    Plotly.toImage(elementID, { format: 'svg', width: 250, height: 50 }).then(function (url) {
         document.getElementById(elementID).src = url;
+        document.getElementById(elementID).classList.remove('image-modify2');
+        document.getElementById(elementID).classList.add('sparklineimages');
     });
 };
 
@@ -2225,7 +2268,7 @@ function disableallbtn() {
     document.getElementById('dropdown-content').disabled = true;
     document.getElementById('maxbtn').disabled = true;
     document.getElementById('minbtn').disabled = true;
-}
+};
 
 function enableallbtn() {
     document.getElementById('prev-btn').disabled = false;
@@ -2263,6 +2306,18 @@ function loading() {
         }
         document.getElementById('lastUpdated').innerHTML = ' ';
         document.getElementById('chart-curvy').innerHTML = '<img src="../images/icons8-loading.gif" class="image-modify2" alt="No data found">';
+        generateSparklineLoading('sparkline-image-voc');
+        generateSparklineLoading('sparkline-image-isc');
+        generateSparklineLoading('sparkline-image-ff');
+        generateSparklineLoading('sparkline-image-vmp');
+        generateSparklineLoading('sparkline-image-imp');
+        generateSparklineLoading('sparkline-image-pmp');
+        generateSparklineLoading('sparkline-image-vop');
+        generateSparklineLoading('sparkline-image-iop');
+        generateSparklineLoading('sparkline-image-pop');
+        generateSparklineLoading('sparkline-image-temperature');
+        generateSparklineLoading('sparkline-image-irradiance');
+        generateSparklineLoading('sparkline-image-perfromance-factor');
 
     } else if (tabStatus === "historic") {
         let labelCurveDiv = document.querySelector('.label-curve-his');
@@ -2276,4 +2331,46 @@ function loading() {
     };
 };
 
+async function getsimulationstatus() {
+    const dataSimulationStatus = { deviceID: sessionStorage.getItem('deviceID') };
+    const optionSimulationStatus = {
+        method: "POST",
+        headers: {
+            "Content-Type": "application/json",
+            'Authorization': `Bearer ${token}`,
+        },
+        body: JSON.stringify(dataSimulationStatus),
+    };
 
+    const responseSimulationStstus = await fetch("/getsimulationStatus", optionSimulationStatus);
+    const simulationresponseData = await responseSimulationStstus.json();
+
+    let foundModelKey = false;
+    let modelValueExists = false;
+
+    for (let item of simulationresponseData) {
+        if (item.key === "model") {
+            foundModelKey = true;
+            // Check if value is present and not null or undefined
+            if (item.value !== null && item.value !== undefined) {
+                modelValueExists = true;
+            }
+            break; // No need to continue once "model" key is found
+        }
+    }
+
+    // Determine simulationStatus based on conditions
+    if (!foundModelKey) {
+        simulationStatus = false; // Key "model" not found
+    } else if (foundModelKey && !modelValueExists) {
+        simulationStatus = false; // Key "model" found but no value
+    } else {
+        simulationStatus = true; // Key "model" found with a valid value
+    }
+};
+
+async function generateSparklineLoading(elementID) {
+    document.getElementById(elementID).classList.remove('sparklineimages');
+    document.getElementById(elementID).src = "../images/icons8-loading.gif";
+    document.getElementById(elementID).classList.add('image-modify2');
+};
